+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/overlays/CustomLocation.kt
@@ -2,14 +2,18 @@ package hamza.dali.flutter_osm_plugin.overlays
 
 import android.graphics.Bitmap
 import android.graphics.Canvas
+import android.graphics.Paint
+import android.graphics.Point
 import android.location.Location
 import hamza.dali.flutter_osm_plugin.VoidCallback
+import hamza.dali.flutter_osm_plugin.utilities.toGeoPoint
 import hamza.dali.flutter_osm_plugin.utilities.toHashMap
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.osmdroid.util.GeoPoint
+import org.osmdroid.util.TileSystem
 import org.osmdroid.views.MapView
 import org.osmdroid.views.Projection
 import org.osmdroid.views.overlay.mylocation.GpsMyLocationProvider
@@ -18,6 +22,8 @@ import org.osmdroid.views.overlay.mylocation.MyLocationNewOverlay
 
 class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
     private var provider: GpsMyLocationProvider? = GpsMyLocationProvider(mapView.context)
+    var disableRotateDirection = false
+    private var mDrawPixel: Point = Point()
 
     init {
         setEnableAutoStop(true)
@@ -40,20 +46,18 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
         disableFollowAndLocation()
         mMyLocationProvider.stopLocationProvider()
     }
-   fun onDestroy(){
-       mMyLocationProvider.destroy()
-   }
-    override fun draw(c: Canvas?, pProjection: Projection?) {
-        mDrawAccuracyEnabled = false
-        super.draw(c, pProjection)
+
+    fun onDestroy() {
+        mMyLocationProvider.destroy()
     }
 
-     fun currentUserPosition(
+
+    fun currentUserPosition(
         result: MethodChannel.Result,
         afterGetLocation: VoidCallback? = null,
         scope: CoroutineScope,
 
-    ) {
+        ) {
         if (!isMyLocationEnabled) {
             enableMyLocation()
         }
@@ -64,7 +68,7 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
                         location.latitude,
                         location.longitude,
                     )
-                    if (!isFollowLocationEnabled){
+                    if (!isFollowLocationEnabled) {
                         disableMyLocation()
                     }
 
@@ -81,10 +85,9 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
 
     override fun onLocationChanged(location: Location?, source: IMyLocationProvider?) {
         super.onLocationChanged(location, source)
-
     }
 
-     fun followLocation(onChangedLocation : (gp:GeoPoint )-> Unit) {
+    fun followLocation(onChangedLocation: (gp: GeoPoint) -> Unit) {
         this.enableFollowLocation()
         runOnFirstFix {
             val location = this.lastFix
@@ -93,23 +96,78 @@ class CustomLocationManager(mapView: MapView) : MyLocationNewOverlay(mapView) {
         }
     }
 
+    override fun draw(canvas: Canvas, pProjection: Projection) {
+        mDrawAccuracyEnabled = false
+        when {
+            disableRotateDirection -> {
+                if (lastFix != null && isMyLocationEnabled) {
+                    drawOnlyPerson(canvas, pProjection, lastFix)
+                }
+            }
+            else -> super.draw(canvas, pProjection)
+        }
+
+    }
+
     fun setMarkerIcon(personIcon: Bitmap?, directionIcon: Bitmap?) {
         when {
             personIcon != null && directionIcon != null -> {
-                setDirectionArrow(
-                    personIcon,
-                    directionIcon
-                )
+                mPersonBitmap = personIcon
+                mDirectionArrowBitmap = directionIcon
+                setDirectionAnchor(.5f, .5f)
                 val mScale = mMapView!!.context.resources.displayMetrics.density
 
-                setPersonHotspot(
-                    mScale * (personIcon.width / 4f) + 0.5f,
-                    mScale * (personIcon.width / 3f) + 0.5f,
-                )
+//                mPersonHotspot.set(
+//                    mScale * (personIcon.width / 4f) + 0.5f,
+//                    mScale * (personIcon.width / 3f) + 0.5f,
+//                )
             }
             personIcon != null -> {
                 setPersonIcon(personIcon)
             }
         }
+    }
+
+    private fun drawOnlyPerson(canvas: Canvas, pProjection: Projection, lastFix: Location) {
+        val mGeoPoint = lastFix.toGeoPoint()
+        pProjection.toPixels(mGeoPoint, mDrawPixel)
+
+        if (mDrawAccuracyEnabled) {
+            val radius: Float = (lastFix.accuracy
+                    / TileSystem.GroundResolution(
+                lastFix.latitude,
+                pProjection.zoomLevel
+            ).toFloat())
+            mCirclePaint.alpha = 50
+            mCirclePaint.style = Paint.Style.FILL
+            canvas.drawCircle(
+                mDrawPixel.x.toFloat(),
+                mDrawPixel.y.toFloat(),
+                radius,
+                mCirclePaint
+            )
+            mCirclePaint.alpha = 150
+            mCirclePaint.style = Paint.Style.STROKE
+            canvas.drawCircle(
+                mDrawPixel.x.toFloat(),
+                mDrawPixel.y.toFloat(),
+                radius,
+                mCirclePaint
+            )
+        }
+        canvas.save()
+        // Unrotate the icon if the maps are rotated so the little man stays upright
+        // Unrotate the icon if the maps are rotated so the little man stays upright
+        canvas.rotate(
+            -mMapView.mapOrientation, mDrawPixel.x.toFloat(),
+            mDrawPixel.y.toFloat()
+        )
+        // Draw the bitmap
+        // Draw the bitmap
+        canvas.drawBitmap(
+            mPersonBitmap, mDrawPixel.x - mPersonHotspot.x,
+            mDrawPixel.y - mPersonHotspot.y, mPaint
+        )
+        canvas.restore()
     }
 }