+++ flutter_osm_web/lib/src/mixin_web.dart
@@ -104,9 +104,13 @@ mixin WebMixin {
     ));
   }
 
-  Future<void> enableTracking({bool enableStopFollow = false}) async {
+  Future<void> enableTracking({
+    bool enableStopFollow = false,
+    bool disableMarkerRotation = false,
+  }) async {
     await interop.enableTracking(
       mapIdMixin,
+      enableStopFollow,
     );
   }
 
@@ -282,6 +286,12 @@ mixin WebMixin {
     roadInfo = roadInfo.copyWith(
       duration: road.duration,
       distance: road.distance,
+      instructions: road.instructions
+          .map((e) => Instruction(
+                instruction: e.instruction,
+                geoPoint: e.location.toGeoPoint(),
+              ))
+          .toList(),
       route: road.polyline!.mapToListGeoPoints(),
     );
     roadsWebCache[roadInfo.key] = roadInfo;